Refactor scenario editor services
