+++ changelog.d/699.changed.md
@@ -1,0 +1,1 @@
+Scenario editor service internals are split by responsibility while preserving the existing public service API and YAML behavior.
